Chore: Move HomeController to src/Presentation/Controllers/Home/
